requirements.txt
boto3
python-dotenv
Pillow
fastapi[all]
uvicorn
clerk-backend-api
requests
python-dotenv
sqlalchemy
africastalking
psycopg2-binary
slowapi
